+++ framework/Error/ErrorHandler.php
@@ -4,6 +4,7 @@ namespace Framework\Error;
 
 use Exception;
 use Framework\Cli\Cli;
+use Framework\Facades\DeveloperTools;
 use Framework\Facades\Env;
 use Framework\Facades\Http;
 use Throwable;
@@ -14,7 +15,7 @@ class ErrorHandler
         int $errno, string $errstr, string $errfile = null, int $errline = null, array $errcontext = null
     ): bool {
         ob_end_clean();
-        if (Env::isDev()) {
+        if (Env::isDev() || DeveloperTools::showErrorMessages()) {
             self::handleDevError($errno, $errstr, $errfile, $errline, $errcontext);
             exit();
         } else {
@@ -26,7 +27,7 @@ class ErrorHandler
     public static function handleException(Throwable $exception): void
     {
         ob_end_clean();
-        if (Env::isDev()) {
+        if (Env::isDev() || DeveloperTools::showErrorMessages()) {
             self::handleDevException($exception);
             exit();
         } else {